Made code more collab request enttity focused abiding by factory design pattern.
