+++ src/main/java/entity/CollabRequest.java
@@ -2,35 +2,14 @@ package entity;
 
 import org.bson.types.ObjectId;
 
-public class CollabRequest {
-    private ObjectId postId;
-    private String author;
-    private String title;
+public interface CollabRequest {
 
-    public CollabRequest() {}
-
-    public CollabRequest(ObjectId postId, String author, String title) {
-        this.postId = postId;
-        this.author = author;
-        this.title = title;
-    }
-    public ObjectId getPostId() {
-        return postId;
-    }
-    public void setPostId(ObjectId postId) {
-        this.postId =  postId;
-    }
-    public String getAuthor() {
-        return author;
-    }
-    public void setAuthor(String author) {
-        this.author = author;
-    }
-    public String getTitle() {
-        return title;
-    }
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
+    ObjectId getPostId();
+    void setPostId(ObjectId postId);
+     String getAuthor();
+   void setAuthor(String author);
+    String getTitle();
+     void setTitle(String title);
+    String getCommenter();
+    void setCommenter(String commenter);
 }
